requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
scikit-learn
sqlalchemy
schedule
loguru
pytest
python-dotenv
